debug: add order

+++ src/main/java/com/se2023/backend/entity/User/User.java
@@ -22,7 +22,7 @@ public class User {
     String confirmCode;
     String role;
     Integer membership;
-    Integer balance;
+    Double balance;
 
 
     public String getUsername() {
